+++ DDR_LocalServer/Servers/LocalServerHeadRuleRouter.cpp
@@ -76,7 +76,7 @@ bool LocalServerHeadRuleRouter::IgnoreBody(std::shared_ptr<BaseSocketContainer> 
 				auto spClientSession = LSClientManager::Instance()->GetTcpClient()->GetConnectedSession();
 				if (spClientSession)
 				{
-					long IntPtr = (long)(spClientSession.get());
+					size_t IntPtr = (size_t)(spClientSession.get());
 					if (passnode.nodetype() == eLocalServer)
 					{
 						if (IntPtr == passnode.receivesessionid())
@@ -93,7 +93,7 @@ bool LocalServerHeadRuleRouter::IgnoreBody(std::shared_ptr<BaseSocketContainer> 
 
 				for (auto spSessionPair : map)
 				{
-					long IntPtr = (long)(spSessionPair.second.get());
+					size_t IntPtr = (size_t)(spSessionPair.second.get());
 					if (passnode.nodetype() == eLocalServer)
 					{
 						if (IntPtr == passnode.receivesessionid())
